requirements.txt
easydict==1.7
h5py==2.10.0
hdf5storage==0.1.15
imageio==2.8.0
imgaug==0.4.0
json-tricks==3.15.2
matplotlib==3.2.1
opencv-python==4.2.0.34
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.2.1
Shapely
six==1.15.0
tensorboardX==1.6
torch==1.0.0
torchvision==0.2.1
yacs==0.1.7
